use explicitly impl instead of new
