+++ src/LibKubernetesGenerator/templates/Operations.cs.template
@@ -6,22 +6,16 @@
 
 namespace k8s;
 
-internal partial class {{name}}Operations : Operations, I{{name}}Operations
+public partial class AbstractKubernetes : I{{name}}Operations
 {
-
-    public {{name}}Operations(AbstractKubernetes kubernetes) : base(kubernetes)
-    {
-    }
-
-
     {{#apis}}
     /// <inheritdoc/>
-    public async Task<HttpOperationResponse{{GetReturnType operation "<>"}}> {{GetMethodName operation "WithHttpMessagesAsync"}}(
+    async Task<HttpOperationResponse{{GetReturnType operation "<>"}}> I{{name}}Operations.{{GetMethodName operation "WithHttpMessagesAsync"}}(
 {{#operation.parameters}}
-        {{GetDotNetType .}} {{GetDotNetName . "true"}},
+        {{GetDotNetType .}} {{GetDotNetName .}},
 {{/operation.parameters}}
-        IReadOnlyDictionary<string, IReadOnlyList<string>> customHeaders = null,
-        CancellationToken cancellationToken = default(CancellationToken))
+        IReadOnlyDictionary<string, IReadOnlyList<string>> customHeaders,
+        CancellationToken cancellationToken)
     {
         var cts = CancellationTokenSource.CreateLinkedTokenSource(cancellationToken);
         cts.CancelAfter(HttpClientTimeout);
